fix: add grid cols
