+++ packages/shared/src/components/FeedItemComponent.tsx
@@ -18,13 +18,14 @@ import { Origin } from '../lib/log';
 import { useFeedLayout, UseVotePost } from '../hooks';
 import { CollectionCard } from './cards/CollectionCard';
 import { CollectionList } from './cards/list/CollectionList';
-import { AcquisitionFormCard } from './cards/AcquisitionFormCard';
 import { MarketingCtaCard } from './marketingCta';
 import { MarketingCtaList } from './marketingCta/MarketingCtaList';
 import { FeedItemType } from './cards/common';
 import { PublicSquadEligibilityCard } from './squads/PublicSquadEligibilityCard';
 import { AdGrid } from './cards/ad/AdGrid';
 import { AdList } from './cards/ad/AdList';
+import { AcquisitionFormGrid } from './cards/AcquisitionForm/AcquisitionFormGrid';
+import { AcquisitionFormList } from './cards/AcquisitionForm/AcquisitionFormList';
 
 const CommentPopup = dynamic(
   () => import(/* webpackChunkName: "commentPopup" */ './cards/CommentPopup'),
@@ -125,6 +126,9 @@ const getTags = ({
     AdTag: useListCards ? AdList : AdGrid,
     PlaceholderTag: useListCards ? PlaceholderList : PlaceholderCard,
     MarketingCtaTag: useListCards ? MarketingCtaList : MarketingCtaCard,
+    AcquisitionFormTag: useListCards
+      ? AcquisitionFormList
+      : AcquisitionFormGrid,
   };
 };
 
@@ -165,7 +169,13 @@ export default function FeedItemComponent({
   );
 
   const { shouldUseListFeedLayout, shouldUseListMode } = useFeedLayout();
-  const { PostTag, AdTag, PlaceholderTag, MarketingCtaTag } = getTags({
+  const {
+    PostTag,
+    AdTag,
+    PlaceholderTag,
+    MarketingCtaTag,
+    AcquisitionFormTag,
+  } = getTags({
     isListFeedLayout: shouldUseListFeedLayout,
     shouldUseListMode,
     postType: (item as PostItem).post?.type,
@@ -248,7 +258,7 @@ export default function FeedItemComponent({
         />
       );
     case FeedItemType.UserAcquisition:
-      return <AcquisitionFormCard key="user-acquisition-card" />;
+      return <AcquisitionFormTag key="user-acquisition-card" />;
     case FeedItemType.PublicSquadEligibility:
       return <PublicSquadEligibilityCard />;
     case FeedItemType.MarketingCta:
