+++ packages/storybook/stories/extension/_providers.tsx
@@ -1,6 +1,6 @@
 import { BootDataProvider } from '@dailydotdev/shared/src/contexts/BootProvider';
 import LogContext from '@dailydotdev/shared/src/contexts/LogContext';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { QueryClient, QueryClientProvider as TanStackQueryClientProvider } from '@tanstack/react-query';
 import { BootApp } from '@dailydotdev/shared/src/lib/boot';
 import { Browser } from 'webextension-polyfill';
 import type { FC, PropsWithChildren } from 'react';
@@ -14,9 +14,13 @@ declare global {
   }
 }
 
+export const QueryClientProvider: FC<PropsWithChildren> = ({ children }) => {
+  return <TanStackQueryClientProvider client={queryClient}>{children}</TanStackQueryClientProvider>;
+}
+
 export const ExtensionProviders: FC<PropsWithChildren> = ({ children }) => {
   return (
-    <QueryClientProvider client={queryClient}>
+    <QueryClientProvider>
       <BootDataProvider
         app={app}
         getRedirectUri={() => `https://daily.dev`}
